[JIT] Support floating point registers in load, spill and copy
We currently want the register class to be sent in as a flag, but that
is actually redundant, since the register class is indicated by the
range of register numbers that a register lies in.

Also, what dynasm calls movd (between a quadword GPR and XMM
register), most everybody else (including intel) calls movq. We might
want to patch dynasm so this is consistent.